nit: create file linera-service/src/cli_wrappers/client_wrapper.rs
